fix: refactor demo account handling and add tests for isDemoAccount utility
